requirements.txt
DB_SERVER=your_server_name
DB_DATABASE=your_database_name
DB_USERNAME=your_username
DB_PASSWORD=your_passwordstreamlit==1.29.0
plotly==5.18.0
pandas==2.1.3
transformers==4.35.2
torch==2.1.1
sqlalchemy
Flask==3.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pymssql==2.2.10
pytest==7.4.3
black==23.11.0
flake8==6.1.0
